+++ bpf/insn/samples/bpf_insn_samples.c
@@ -9,14 +9,9 @@ struct bpf_insn_sample {
 #define ARRAY_SIZE(arr)	(sizeof(arr) / sizeof(arr[0]))
 
 static struct bpf_insn_sample samples[] = {
-#define SAMPLE(sample) { .name = #sample, .get_insns = BPF_INSN_SAMPLE_FUNC_NAME(sample), }
-	SAMPLE(trace_printk),
-	SAMPLE(strncmp),
-	SAMPLE(loop),
-	SAMPLE(cgroup_from_id),
-	SAMPLE(task_from_pid),
-	SAMPLE(cgrp_storage_get),
-	SAMPLE(get_func_ip),
+#define BPF_INSN_SAMPLE(sample) { .name = #sample, .get_insns = BPF_INSN_SAMPLE_FUNC_NAME(sample), },
+#include "bpf_insn_defs.h"
+#undef BPF_INSN_SAMPLE
 };
 
 bpf_samples_get_insns_fn bpf_samples_get_insns_from_string(const char *helper_str)